fastapi/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic-settings
python-multipart
spacy
